refactor(misc): :zap: removed ParsingCandidate, RawHelper, URLHelper classes. Removed link_hash from article (was never used)
article.link_hash is no longer available
