+++ newspaper/languages/__init__.py
@@ -228,8 +228,15 @@ def get_language_from_iso639_1(iso639_1: str) -> Optional[str]:
     return languages_dict.get(iso639_1)
 
 
+def get_available_languages():
+    """Returns a list of available languages and their 2 char input codes"""
+    stopword_files = Path(settings.STOPWORDS_DIR).glob("stopwords-??.txt")
+    for file in stopword_files:
+        yield file.stem.split("-")[1]
+
+
 def valid_languages():
-    """Returns the List of available Languages"""
+    """Returns the List of available Languages as tuples (iso-code, language)"""
     languages = []
     for code, language in languages_tuples:
         stopwords_file = Path(settings.STOPWORDS_DIR) / f"stopwords-{language}.txt"
